refactor: rename stripe modules
