+++ storefronts/features/checkout/gateways/stripeGateway.js
